[fix] Displacement::position() should not return invalid positions
When travelling on an horizontal line, the speed vector is (2xt,0).
With the old position() method; the position at t=0.5 would be (1,0),
which is invalid in the pitch coordinate system.

The new algorithm use a special case: if we travel on a horizontal line,
ensure that we are on an hexagon

..Sarah Cox..